move transfer code to a service. Add background job. Connect command to service
